+++ Libplanet.Net/Protocols/KademliaProtocol.cs
@@ -61,7 +61,7 @@ namespace Libplanet.Net.Protocols
         // when found neighbor's count is reached 2*k.
         public async Task BootstrapAsync(
             IEnumerable<BoundPeer> bootstrapPeers,
-            TimeSpan? pingSeedTimeout,
+            TimeSpan? dialTimeout,
             int depth,
             CancellationToken cancellationToken)
         {
@@ -74,7 +74,7 @@ namespace Libplanet.Net.Protocols
                 // Guarantees at least one connection (seed peer)
                 try
                 {
-                    await PingAsync(peer, pingSeedTimeout, cancellationToken);
+                    await PingAsync(peer, dialTimeout, cancellationToken);
                     findPeerTasks.Add(
                         FindPeerAsync(
                             history,
@@ -82,7 +82,7 @@ namespace Libplanet.Net.Protocols
                             _address,
                             peer,
                             depth,
-                            pingSeedTimeout,
+                            dialTimeout,
                             cancellationToken));
                 }
                 catch (PingTimeoutException)